+++ slack-strava/commands/help.rb
@@ -17,6 +17,7 @@ module SlackStrava
         Settings
         --------
         set units mi|km          - use miles or kilometers
+        set fields all|none|...  - display all, none or certain activity fields
         set maps off|full|thumb  - change the way maps are displayed
         set private true|false   - sync private activities (default is false)