+++ slack-strava/models.rb
@@ -1,5 +1,7 @@
 require 'slack-strava/models/error'
 require 'slack-strava/models/brag'
+require 'slack-strava/models/array_validator'
+require 'slack-strava/models/activity_fields'
 require 'slack-strava/models/team'
 require 'slack-strava/models/athlete'
 require 'slack-strava/models/user'