+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 2018/6/05: Added `set fields` to choose which fields to display - [@dblock](https://github.com/dblock).
 * 2018/6/05: Display both speed and pace - [@dblock](https://github.com/dblock).
 * 2018/5/28: Update activity labels according to the strava API - [@gmiossec](https://github.com/gmiossec).
 * 2018/5/28: Ride activities pace is displayed in km/h if units is km - [@gmiossec](https://github.com/gmiossec).